+++ thunderstore/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Shaman Dive Bind 
 
+## v1.1.0 
+- Added Volt Filament particles and audio to the dive shockwave when equipped 
+- Volt Filament now recolors the dive sprites when equipped 
+- Added glow/bloom effect to the dive shockwave 
+- Reworked how the Shaman Crest damage multiplier is applied to the damaging hitboxes internally. The effect in game should be identical to previous versions, but should now work properly for custom crests using the Shaman Crest bind 
+
 ## v1.0.1 
 - Fixed shockwave hitbox being enabled on all crests when binding